// src/slices/types.ts
export type SerializedBigNumber = string | null;

interface SerializedFarmUserData {
  allowance: string;
  tokenBalance: string;
  stakedBalance: string;
  earnings: string;
}

export interface DeserializedFarmUserData {
  allowance: BigNumber;
  tokenBalance: BigNumber;
  stakedBalance: BigNumber;
  earnings: BigNumber;
}

export interface SerializedFarm extends SerializedFarmConfig {
  tokenPriceBusd?: string;
  quoteTokenPriceBusd?: string;
  tokenAmountTotal?: SerializedBigNumber;
  quoteTokenAmountTotal?: SerializedBigNumber;
  lpTotalInQuoteToken?: SerializedBigNumber;
  lpTotalSupply?: SerializedBigNumber;
  tokenPriceVsQuote?: SerializedBigNumber;
  poolWeight?: SerializedBigNumber;
  userData?: SerializedFarmUserData;
}

export interface DeserializedFarm extends DeserializedFarmConfig {
  tokenPriceBusd?: string;
  quoteTokenPriceBusd?: string;
  tokenAmountTotal?: BigNumber;
  quoteTokenAmountTotal?: BigNumber;
  lpTotalInQuoteToken?: BigNumber;
  lpTotalSupply?: BigNumber;
  tokenPriceVsQuote?: BigNumber;
  poolWeight?: BigNumber;
  userData?: DeserializedFarmUserData;
}

export enum VaultKey {
  CakeVault = "cakeVault",
}

interface CorePoolProps {
  startBlock?: number;
  endBlock?: number;
  apr?: number;
  rawApr?: number;
  stakingTokenPrice?: number;
  earningTokenPrice?: number;
  vaultKey?: VaultKey;
}

export interface DeserializedPool extends DeserializedPoolConfig, CorePoolProps {
  totalStaked?: BigNumber;
  stakingLimit?: BigNumber;
  stakingLimitEndBlock?: number;
  profileRequirement?: {
    required: boolean;
    thresholdPoints: BigNumber;
  };
  userData?: {
    allowance: BigNumber;
    stakingTokenBalance: BigNumber;
    stakedBalance: BigNumber;
    pendingReward: BigNumber;
  };
}

export interface SerializedPool extends SerializedPoolConfig, CorePoolProps {
  totalStaked?: SerializedBigNumber;
  stakingLimit?: SerializedBigNumber;
  numberBlocksForUserLimit?: number;
  profileRequirement?: {
    required: boolean;
    thresholdPoints: SerializedBigNumber;
  };
  userData?: {
    allowance: SerializedBigNumber;
    stakingTokenBalance: SerializedBigNumber;
    stakedBalance: SerializedBigNumber;
    pendingReward: SerializedBigNumber;
  };
}
// ...
export interface SerializedVaultFees {
  performanceFee: number | null;
  withdrawalFee: number | null;
  withdrawalFeePeriod: number | null;
}

export interface DeserializedVaultFees extends SerializedVaultFees {
  performanceFeeAsDecimal: number;
}

interface SerializedVaultUser {
  isLoading: boolean | null;
  userShares: SerializedBigNumber | null;
  cakeAtLastUserAction: SerializedBigNumber | null;
  lastDepositedTime: string | null;
  lastUserActionTime: string | null;
}

export interface SerializedLockedVaultUser extends SerializedVaultUser {
  lockStartTime: string | null;
  lockEndTime: string | null;
  userBoostedShare: SerializedBigNumber | null;
  locked: boolean | null;
  lockedAmount: SerializedBigNumber | null;
  currentPerformanceFee: SerializedBigNumber | null;
  currentOverdueFee: SerializedBigNumber | null;
}

export interface DeserializedVaultUser {
  isLoading: boolean | null;
  userShares: BigNumber | null;
  cakeAtLastUserAction: BigNumber | null;
  lastDepositedTime: string | null;
  lastUserActionTime: string | null;
}

export interface DeserializedLockedVaultUser extends DeserializedVaultUser {
  lastDepositedTime: string | null;
  lastUserActionTime: string | null;
  lockStartTime: string | null;
  lockEndTime: string | null;
  userBoostedShare: BigNumber | null;
  locked: boolean | null;
  lockedAmount: BigNumber | null;
  balance: {
    cakeAsNumberBalance: number | null;
    cakeAsBigNumber: BigNumber | null;
    cakeAsDisplayBalance: string | null;
  };
  currentPerformanceFee: BigNumber | null;
  currentOverdueFee: BigNumber | null;
}

export interface DeserializedIfoVaultUser extends DeserializedVaultUser {
  credit: string;
}

export interface DeserializedCakeVault {
  totalShares?: BigNumber;
  totalLockedAmount?: BigNumber;
  pricePerFullShare?: BigNumber;
  totalCakeInVault?: BigNumber;
  fees?: DeserializedVaultFees;
  userData?: DeserializedLockedVaultUser;
}

export interface SerializedCakeVault {
  totalShares?: SerializedBigNumber | null;
  totalLockedAmount?: SerializedBigNumber | null;
  pricePerFullShare?: SerializedBigNumber | null;
  totalCakeInVault?: SerializedBigNumber | null;
  fees?: SerializedVaultFees | null;
  userData?: SerializedLockedVaultUser | null;
}